fix(input): move Input from universal core to client core
