+++ src/components/client/core/index.ts
@@ -4,6 +4,7 @@ export { default as Button } from "./Button/Button";
 export { default as Checkbox } from "./Checkbox/Checkbox";
 export { default as Drawer } from "./Drawer/Drawer";
 export { default as Icon } from "./Icon/Icon";
+export { default as Input } from "./Input/Input";
 export { default as Menu } from "./Menu/Menu";
 export { default as Modal } from "./Modal/Modal";
 export { default as Slider } from "./Slider/Slider";
@@ -19,6 +20,7 @@ export type { Props as ButtonProps } from "./Button/Button";
 export type { Props as CheckboxProps } from "./Checkbox/Checkbox";
 export type { Props as DrawerProps } from "./Drawer/Drawer";
 export type { Props as IconProps } from "./Icon/Icon";
+export type { Props as InputProps } from "./Input/Input";
 export type {
   Props as MenuProps,
   MenuItemRecord,
